requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
python-pptx>=0.6.21
streamlit>=1.40.0
openai>=1.50.0
httpx>=0.27.0
